day 1
+ resources done
+ basic resource upgrades done
+ basic player pawn (camera stuff) done

next:
player/ai base
sending out grunts
initial gold for leaving town
bounties

+++ Source/Growth/Public/Components/ResourceComponent.h
@@ -96,13 +96,13 @@ public:
 	}
 
 	UFUNCTION(BlueprintCallable, BlueprintPure, Category = "Resources")
-	float GetCurrentSupply()
+	int32 GetCurrentSupply()
 	{
 		return CurrentSupply;
 	}
 	
 	UFUNCTION(BlueprintCallable, BlueprintPure, Category = "Resources")
-	float GetMaximumSupply()
+	int32 GetMaximumSupply()
 	{
 		return MaximumSupply;
 	}
